Modified the compare_checks function in ExecutableSequence to handle sequences that create different observers on different runs of the sequence.  This can happen because of changes in the environment (eg, static variables) between runs.  Some observations are only created when the value involved in the observation meets certain criteria and since the values can change, the observations may not always be created.  Its also possible to throw exceptions in one run and not another.
The new version uses a string id (get_id() from the Check
interface) to identify an observation uniquely.  The id contains
both the observer type (primitive, non-null, observer method call,
etc) and the related variables.  When comparing two lists of
Checks, the code presumes that the checks will occur in the same
order on each execution.  Thus if the first remaining check in
the first execution matches the second check from the second execution,
the first check from the second execution is presumed to not match.
The algorithm is far from completely optimized, but it should work
pretty well if the two lists are pretty much the same (which
they usually are).

This required adding get_observer_str() to the ObjectContract
interface.  This returns the type of the observer.  This is
used by ObjectCheck to build the string id.