Create new test check visitors
Moved test generation from ExecutionVisitor implementors to implementors
of TestCheckGenerator, which is also a visitor pattern, but returns a
TestCheck object.
ContractCheckingVisitor is modified to use new interface and to use
predicates that determine whether an exception is to be considered an error.
RegressionCaptureVisitor is also modified to use predicates to change
behavior based on command line arguments.
The GenerateBoth and GenerateRegressionOnly visitors combine the ContractCheckingVisitor and RegressionCaptureVisitor, and the DummyCheckGenerator returns
an empty RegressionChecks object.