Fix short format in methods and sequences
Fix short format casting in method call so that autoboxed primitive receiver
is possible, and so that casting handled properly when expected and actual
type the same.
Fix sequence so that doesn't do short format when statement is only or last
statement in sequence.
Fix texts that expect long format output.