Move test driver logic to test classes.
TestDriver was generating gigantic methods that won't compile.

My first attempt to solve this was to split the driver's method in
one method per test class. This didn't work. Sometimes the driver's
class is too big, and `javac` gives a `too many constants` error.

This commits creates a static `testAll` method in each test class,
so the driver only has to invoke them.